% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{allocate_areas_survey_regions}
\alias{allocate_areas_survey_regions}
\title{Allocate areas to survey regions}
\usage{
allocate_areas_survey_regions(areas_wide, survey_region_boundaries)
}
\arguments{
\item{areas_wide}{wide format area hierarchy, created by \code{naomi::spread_areas()}.}

\item{survey_region_boundaries}{survey_region_boundaries dataset created by
\code{create_survey_boundaries_dhs()}.}
}
\value{
A simple features data frame consisting of a mapping of all areas
to a survey_region_id.
}
\description{
Allocate areas at the most granular level to survey regions via spatial
join based on largest overlapping area.
}
\details{
The function \code{sf:st_join(..., largest = TRUE)} is used to construct a spatial
join based on the area of largest overlap.

If the mapping is clean, the following should be satisfied:
\enumerate{
\item All areas are allocated to a survey region. This might not happen if
an area is non-overlapping with the survey geometry.
\item All survey regions should contain some areas. This might not happen if
all areas overlapping a region are not cleanly nested and have a
larger overlap with other regions.
}

The function \code{assert_survey_region_areas()} implements these checks.

These conditions are not comprehensive and do not guarantee the mapping is
accurate, but will catch some basic errors.
}
