% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{assert_area_id_check}
\alias{assert_area_id_check}
\title{Checks for consistent area IDs between two datasets}
\usage{
assert_area_id_check(df1, df2, key)
}
\arguments{
\item{df1}{a dataframe containing area_id}

\item{df2}{a dataframe containing area_id}

\item{key}{list of columns to compare}
}
\value{
If unique area IDs are present between the
two datasets, an error will be generated along
with a map of mismatching IDs
}
\description{
Checks for consistent area IDs between two datasets
}
