% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{assert_pop_age_group}
\alias{assert_pop_age_group}
\title{Checks valid age groups}
\usage{
assert_pop_age_group(var)
}
\arguments{
\item{var}{a value in dataframe extracted using \code{$} notation}
}
\value{
If additional age groups are present or missing values
for age group, an error will be generated
}
\description{
Checks valid age groups
}
