% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{assert_pop_data_check}
\alias{assert_pop_data_check}
\title{Data validation for input population data}
\usage{
assert_pop_data_check(pop_data, boundaries)
}
\arguments{
\item{pop_data}{population dataframe}

\item{boundaries}{boundary file conatining area IDs}
}
\value{
If unique area IDs are present between the
two datasets, an error will be generated along
with a map of mismatching IDs. If unique age groups area present,
an error will be generated.
}
\description{
Checks for:
(1) consistent aread IDs between pop data and boundaries file
(2) pop data age groups consistent with naomi age groups
}
