% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{assert_shipp_resource_hierarchy}
\alias{assert_shipp_resource_hierarchy}
\title{Throw warning when area hierarchy in external SHIPP resources read in from
\code{naomi.resources} do not match Naomi outputs used to update SHIPP estimates.}
\usage{
assert_shipp_resource_hierarchy(outputs, options)
}
\arguments{
\item{outputs}{Naomi outputs.}

\item{options}{Naomi options.}
}
\value{
Resource hierarchy
}
\description{
Throw warning when area hierarchy in external SHIPP resources read in from
\code{naomi.resources} do not match Naomi outputs used to update SHIPP estimates.
}
\keyword{internal}
