% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_survey_indicators.R
\name{calc_all_outcomes}
\alias{calc_all_outcomes}
\title{Calculate ...}
\usage{
calc_all_outcomes(
  ind,
  group_by_vars,
  split_vars,
  formula,
  extra_vars,
  age_groups
)
}
\description{
Calculate ...
}
