% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pjnz.R
\name{check_pjnz_shiny90}
\alias{check_pjnz_shiny90}
\title{Check whether PJNZ contains .shiny90 file}
\usage{
check_pjnz_shiny90(pjnz)
}
\arguments{
\item{pjnz}{file path to PJNZ}
}
\value{
Logical whether PJNZ file contains a .shiny90 file
}
\description{
Check whether PJNZ contains .shiny90 file
}
\details{
TODO: Check whether the .shiny90 file is valid.
}
