% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_phia_sexbehav.R
\name{check_survey_sexbehav}
\alias{check_survey_sexbehav}
\title{Check that each individual is assigned to one and only one risk category.}
\usage{
check_survey_sexbehav(survey_sexbehav)
}
\arguments{
\item{survey_sexbehav}{Output of e.g. \code{extract_sexbehav_phia}.}
}
\value{
Rows of \code{survey_sexbehav} not in exactly one category.
}
\description{
Check that each individual is assigned to one and only one risk category.
}
