% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pjnz.R
\name{copy_pjnz_extract}
\alias{copy_pjnz_extract}
\title{Extract the .DP and .PJN from a Spectrum PJNZ}
\usage{
copy_pjnz_extract(pjnz, out, shiny90 = NULL, force_shiny90 = FALSE)
}
\arguments{
\item{pjnz}{file path to source PJNZ}

\item{out}{file path to save output}

\item{shiny90}{file path to external .shiny90 zip (optional)}

\item{force_shiny90}{Logical whether or not to force replacement
of a .shiny90 file already in the PJNZ with the provided path.
The default behaviour is not to replace the .shiny90 file
if it already exists in the PJNZ.}
}
\description{
Copy a PJNZ file to a new location an delete everything except for the
.DP and .PJN files.
}
\details{
Both pjnz and out must be length 1. To apply to multiple files, use
\code{Map} function, e.g. \code{Map(copy_pjnz_extract, pjnz_list, out_list)}.

The file must be renamed (pjnz cannot equal out) to avoid inadvertently
deleting components from an archived PJNZ file.

The default `force_shiny90 = FLASE)
}
