% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs_hts.R
\name{create_hivtesting_dhs}
\alias{create_hivtesting_dhs}
\title{Create individual HIV testing history variables from DHS}
\usage{
create_hivtesting_dhs(surveys)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}
}
\value{
data.frame consisting of survey ID, cluster ID and individual
HIV testing history outcomes. See details.
}
\description{
Create dataset of individual HIV testing history variables.
}
\details{
The following fields are extracted:
\itemize{
\item survey_id
\item individual_id
\item evertest
\item test12m
}

Variable \code{evertest} is the outcome ever been tested for HIV and received
the result of the most recent HIV test. Variable \code{test12m} is whether
the individual has been tested for HIV and received the result in the
past 12 months.
}
\examples{

surveys <- create_surveys_dhs("MWI")
hts <- create_hivtesting_dhs(surveys)

}
