% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_individual_dhs}
\alias{create_individual_dhs}
\title{Create individual non-HIV outcomes dataset from DHS}
\usage{
create_individual_dhs(surveys)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}
}
\value{
data.frame consisting of survey ID, cluster ID and individual
demographic outcomes. See details.
}
\description{
Create dataset of indiviaul demographic and HIV outcomes. For use with DHS surveys
that don't include HIV biomarker testing. Currently outputs all same variables
as create_individual_hiv_dhs for consistency to feed into calc_survey_indicators
}
\details{
The following fields are extracted:
\itemize{
\item survey_id
\item cluster_id
\item household
\item line
\item sex
\item age
\item dob_cmc
\item interview_cmc
\item indweight
\item hivstatus (NA)
\item arv (NA)
\item artself (NA)
\item vls (NA)
\item cd4 (NA)
\item artall (NA)
\item hivweight (NA)
\item recent (NA)
}
}
\examples{

surveys <- create_surveys_dhs("MWI")
individuals <- create_individual_dhs(surveys)

}
