% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_survey_boundaries_dhs}
\alias{create_survey_boundaries_dhs}
\title{Create survey region boundaries dataset from DHS spatial data repository}
\usage{
create_survey_boundaries_dhs(surveys)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}
}
\value{
A simple features data frame containing DHS region code, region name,
and region boundaries for each survey.
}
\description{
Create survey region boundaries dataset from DHS spatial data repository
}
\details{
For some surveys, the DHS spatial data repository and the survey clusters
datasets boundaries at multiple levels (e.g. admin 1 and admin 2). In these
cases, the admin level with the largest number or regions is selected by
default.

TODO: Support a workflow for selecting alternative levels.
}
