% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_survey_individuals_dhs}
\alias{create_survey_individuals_dhs}
\alias{create_survey_biomarker_dhs}
\title{Create survey individuals and biomarker dataset from DHS extract}
\usage{
create_survey_individuals_dhs(dat)

create_survey_biomarker_dhs(dat)
}
\arguments{
\item{dat}{data.frame of merged individual extract, returned by
\code{create_individual_hiv_dhs()}.}
}
\value{
data.frame matching UNAIDS data schema
}
\description{
Create survey individuals and biomarker dataset from DHS extract
}
