% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_survey_regions_dhs}
\alias{create_survey_regions_dhs}
\title{Create survey regions dataset from DHS}
\usage{
create_survey_regions_dhs(survey_region_areas)
}
\arguments{
\item{survey_region_areas}{Area allocation to survey regions, created by
\code{\link[=allocate_areas_survey_regions]{allocate_areas_survey_regions()}}}
}
\value{
Survey regions dataset conforming to schema.
}
\description{
Construct survey regions dataset by identifying the smallest area_id
that contains the whole survey region.
}
