% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs_hts.R
\name{extract_hivtesting_dhs}
\alias{extract_hivtesting_dhs}
\title{Recode v826a to v826
c("less than 12 months"=1, "12-23 months"=2, "2 years or more"=3),}
\usage{
extract_hivtesting_dhs(SurveyId, ird_path, mrd_path)
}
\description{
Recode v826a to v826
c("less than 12 months"=1, "12-23 months"=2, "2 years or more"=3),
}
