% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_mics_sexbehav.R
\name{extract_sexbehav_mics}
\alias{extract_sexbehav_mics}
\title{Extract sexual behaviour categorical variables from PHIA surveys.}
\usage{
extract_sexbehav_mics(ind, survey_id, gender)
}
\arguments{
\item{ind}{Individuals dataset.}

\item{survey_id}{The survey name.}
}
\value{
Sexual behaviour categorical variables
}
\description{
Extract sexual behaviour categorical variables from PHIA surveys.
}
