% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{generate_area_id}
\alias{generate_area_id}
\title{Generate single Naomi area id}
\usage{
generate_area_id(iso3, level, nchar = 5)
}
\arguments{
\item{iso3}{three character ISO3 code}

\item{level}{area level as an integer}

\item{nchar}{number of alpha numeric digits to generate}
}
\value{
An area_id in the format \verb{<ISO3>_<level>_<xyz12>}.
}
\description{
Generate a Naomi area ID consisting of ISO3, area level and
a random \code{nchar} digit alpha numeric.
}
\details{
This function is not vectorized. It generates a single area ID.

This function does not set the seed. Ensure to set the seed
before calling the function if you want to reproduce the
same results.
}
\examples{

generate_area_id("ISO", 1)

}
