% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{hintr_inputs_ready}
\alias{hintr_inputs_ready}
\title{Prepare output from hintr debug rds for debugging}
\usage{
hintr_inputs_ready(jobid, root = ".")
}
\arguments{
\item{jobid}{The model fit or calibrate ID (folder created by \code{hintr::download_debug()})}

\item{root}{The ticket folder, i.e. the path returned by \code{naomi_debug()}'s \code{dest}}
}
\value{
Path to local debug
}
\description{
Prepare output from hintr debug rds for debugging
}
