% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_population.R
\name{naomi_extract_worldpop}
\alias{naomi_extract_worldpop}
\title{Extract WorldPop raster data}
\usage{
naomi_extract_worldpop(
  areas,
  iso3 = areas$area_id[areas$area_level == 0],
  years = c(2010, 2015, 2020)
)
}
\arguments{
\item{areas}{Naomi area hierarchy dataset with boundaries.}

\item{iso3}{ISO3 country code.}

\item{years}{Years to extract WorldPop data}
}
\value{
A data frame formatted as Naomi population dataset
}
\description{
Extract WorldPop raster data
}
\details{
Raster files are downloaded from the WorldPop FTP. Some files are very large.
It is recommended to run this on a fast internet connection.
}
