% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{plot_area_hierarchy_summary}
\alias{plot_area_hierarchy_summary}
\title{Plot area hierarchy levels}
\usage{
plot_area_hierarchy_summary(areas, nrow = 1)
}
\arguments{
\item{areas}{area hierarchy sf object}

\item{nrow}{number of rows, integer.}
}
\value{
A ggplot2 object illustrating the area hierarchy
}
\description{
Plot area hierarchy levels
}
