% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pjnz.R
\name{read_pjnz_region_code}
\alias{read_pjnz_region_code}
\title{Read Spectrum region code from PJNZ file}
\usage{
read_pjnz_region_code(pjnz)
}
\arguments{
\item{pjnz}{file path to source PJNZ}
}
\description{
Read Spectrum region code from PJNZ file
}
