% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{read_sf_zip}
\alias{read_sf_zip}
\title{Read shape file from ZIP}
\usage{
read_sf_zip(zfile, pattern = "shp$")
}
\arguments{
\item{zfile}{Path to zip file}

\item{pattern}{Pattern to read files for from zip, defaults to files ending
with 'shp'}
}
\description{
Read shape file from ZIP
}
