% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-naomi1-naomi2.R
\name{recode_naomi1_age_group}
\alias{recode_naomi1_age_group}
\title{Recode age group from Naomi 1 to Naomi 2}
\usage{
recode_naomi1_age_group(x)
}
\arguments{
\item{x}{Character vector of age groups in Naomi 1 format}
}
\value{
Character vector of age groups in Naomi 2 format
}
\description{
Recode age group from Naomi 1 to Naomi 2
}
\examples{
recode_age_group(c("15-19", "15+", "00+"))

}
