% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-naomi1-naomi2.R
\name{recode_naomi1_art}
\alias{recode_naomi1_art}
\alias{recode_naomi1_anc}
\title{Update ART and ANC programme data set to Naomi 2.0 specifications}
\usage{
recode_naomi1_art(art)

recode_naomi1_anc(anc)
}
\arguments{
\item{art}{Data frame of ART data conforming to Naomi 1.0 schema.}

\item{anc}{Data frame of ANC testing data conforming to Naomi 1.0 schema.}
}
\value{
Data frame of ART data conforming to Naomi 2.0 schema.
}
\description{
Update ART and ANC programme data set to Naomi 2.0 specifications
}
\details{
\itemize{
\item Rename \code{current_art} column to \code{art_current}.
\item Recode \code{year} column to \code{calendar_quarter} in ART dataset.
\item Recode \code{age_group} column from \code{15-49} format to \code{Y015_049}.
\item Recode \verb{ancrt_*} columns to \verb{anc_*}.
}
}
