% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_calculate_incidence_female}
\alias{shipp_calculate_incidence_female}
\title{Calculate incidence for female SRB groups.}
\usage{
shipp_calculate_incidence_female(
  naomi_output,
  options,
  iso,
  female_srb,
  female_logit_prevalence,
  survey_year,
  consensus_est,
  scale_fsw_new_infections,
  goals,
  kp_wb
)
}
\arguments{
\item{naomi_output}{Naomi output.}

\item{options}{Naomi model options.}

\item{female_srb}{FSW adjusted estimates of female sexual risk groups generated by \code{shipp_adjust_sexbehav_fsw()}.}

\item{female_logit_prevalence}{Risk adjusted estimates of female prevalence in sexual risk groups generated by \code{shipp_calculate_prevalence_female()}.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}

\item{consensus_est}{Source to scale PSE + new infections.}

\item{goals}{National PSE and new infection estimates from Goals.}

\item{kp_wb}{National PSE and new infection estimates from KP Workbook read from pjnz.}
}
\value{
Wide format output required for the SHIPP workbook.
}
\description{
While maintaining age/sex/district-specific HIV incidence from Naomi, distribute
HIV incidence between our 4 different behavioural groups utilizing IRRs from the
literature
}
\keyword{internal}
