% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_combine_cats_female}
\alias{shipp_combine_cats_female}
\title{Prep women's outputs for spreadsheet tabs}
\usage{
shipp_combine_cats_female(age_filter, shipp, naomi_output)
}
\arguments{
\item{age_filter}{formatted as eg "Y015_019"}

\item{shipp}{female outputs}

\item{naomi_output}{naomi output used for pulling in country & area names}
}
\value{
dataframe ready to write to Excel tab
}
\description{
Prep women's outputs for spreadsheet tabs
}
