% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_combine_cats_male}
\alias{shipp_combine_cats_male}
\title{Prep men's outputs for spreadsheet tabs, including combining non-regular
partner group with KPs to make "elevated risk" group}
\usage{
shipp_combine_cats_male(age_filter, shipp, naomi_output)
}
\arguments{
\item{age_filter}{formatted as eg "Y015_019"}

\item{shipp}{male outputs}

\item{naomi_output}{naomi output used for pulling in country & area names}
}
\value{
dataframe ready to write to Excel tab
}
\description{
Prep men's outputs for spreadsheet tabs, including combining non-regular
partner group with KPs to make "elevated risk" group
}
