% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_generate_risk_populations}
\alias{shipp_generate_risk_populations}
\title{Generate outputs to update SHIPP tool.}
\usage{
shipp_generate_risk_populations(
  naomi_output,
  pjnz = NULL,
  consensus_est,
  scale_fsw_new_infections,
  scale_msm_new_infections = TRUE,
  scale_pwid_new_infections = TRUE,
  scale_fsw_pse = TRUE,
  scale_msm_pse = TRUE,
  scale_pwid_pse = TRUE
)
}
\arguments{
\item{naomi_output}{Path to naomi output (zip file or hintr object).}

\item{pjnz}{Path to spectrum file.}

\item{consensus_est}{Source of consensus estimate to scale PSE and KP new infections.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}
}
\value{
Output files to update SHIPP excel workbook.
}
\description{
Generate outputs to update SHIPP tool.
}
\keyword{internal}
