% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_reformat_output}
\alias{shipp_reformat_output}
\title{Reformat output for individual spreadsheet tabs (e.g. "PSE F15-19")}
\usage{
shipp_reformat_output(shipp_all)
}
\arguments{
\item{shipp_all}{list of SHIPP outputs from shipp_generate_risk_populations}
}
\value{
list of output for each spreadsheet tab
}
\description{
Reformat output for individual spreadsheet tabs (e.g. "PSE F15-19")
}
