% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{surveys_add_dhs_regvar}
\alias{surveys_add_dhs_regvar}
\title{Add REGVAR to surveys dataset}
\usage{
surveys_add_dhs_regvar(surveys, survey_region_boundaries)
}
\arguments{
\item{surveys}{surveys dataset, data.frame.}

\item{survey_region_boundaries}{survey_region_boundaries dataset, sf object.}
}
\value{
The surveys data.frame
}
\description{
The variable name for the survey region variable is sourced from the
DHS survey boundaries datasets sourced by \code{create_survey_boundaries_dhs()}.
Utility function to merge survey region variable name to \code{surveys}
dataset from \code{survey_region_boundaries} dataset.
}
\details{
This will throw an error if the REGVAR is not unique to each survey_id within
the \code{survey_region_boundaries} dataset.
}
