% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_areas.R
\name{write_sf_shp_zip}
\alias{write_sf_shp_zip}
\title{Save sf object to zipped ESRI .shp file}
\usage{
write_sf_shp_zip(obj, zipfile, overwrite = FALSE)
}
\arguments{
\item{obj}{an object of class \code{sf}.}

\item{zipfile}{path to write zip output file. Must have file extension .zip.}

\item{overwrite}{logical whether to overwrite \code{zipfile} if it already exists.}
}
\value{
Return value of \code{file.copy()}, \code{TRUE} if file successfully written.
}
\description{
Save an sf object as a zipped archive with the four ESRI shape
file components \code{.shp}, \code{.dbf}, \code{.prj}, \code{.shx}. This wraps
\code{\link[sf:st_write]{sf::write_sf()}}.
}
\examples{
nc <- sf::read_sf(system.file("shape/nc.shp", package="sf"))
write_sf_shp_zip(nc, "nc.zip")

}
