% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_individual_hiv_dhs}
\alias{create_individual_hiv_dhs}
\title{Create individual HIV outcomes dataset from DHS}
\usage{
create_individual_hiv_dhs(surveys, clear_rdhs_cache = FALSE)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}

\item{clear_rdhs_cache}{Passed to \code{rdhs::get_datasets(clear_cache = )}.}
}
\value{
data.frame consisting of survey ID, cluster ID and individual
demographic and HIV outcomes. See details.
}
\description{
Create dataset of indiviaul demographic and HIV outcomes.
}
\details{
The following fields are extracted:
\itemize{
\item survey_id
\item cluster_id
\item household
\item line
\item sex
\item age
\item dob_cmc
\item interview_cmc
\item indweight
\item hivstatus
\item arv
\item artself
\item vls
\item cd4
\item artall
\item hivweight
}
}
\examples{
\dontrun{
surveys <- create_surveys_dhs("MWI")
individuals <- create_individual_hiv_dhs(surveys)
}

}
