% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs_sexbehav.R
\name{create_sexbehav_dhs}
\alias{create_sexbehav_dhs}
\title{Create individual sexual behaviour variables from DHS}
\usage{
create_sexbehav_dhs(surveys)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}
}
\value{
data.frame consisting of survey ID, cluster ID and individual
sexual behaviour outcomes. See details.
}
\description{
Create dataset of individual sexual behaviour variables.
}
\details{
The following fields are extracted:
\itemize{
\item survey_id
\item individual_id
\item eversex
\item sex12m
\item sexcohab
\item sexnonreg
\item sexpaid12mo
\item sti12mo
}

Variable \code{eversex} is the outcome ever reporting sexual activity.
Variable \code{sex12m} is whether an individual reports having been sexually
active in the past 12 months. Variable \code{sexcohab} is whether an individual
reports being sexually active in the past 12 months with only one cohabiting
partner. Variable \code{sexnonreg} is whether the individual reports having
non-regular sexual partner(s) or multiple partners in the past year. Variable
\code{sexpaid12m} for females is whether respondent reports having had sex in return
for cash, gifts, or anything else in the past 12 months (only asked of 15-24
y/o women in DHS7) and for males is whether respondent reports having paid
for sex in the past 12 months. Variable \code{sti12m} is whether a respondent
reports having any STD, genital sore/ulcer, or genital discharge in the last
12 months.
}
\examples{

surveys <- create_surveys_dhs("MWI")
sexbehav <- create_sexbehav_dhs(surveys)

}
