% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_survey_circumcision_dhs}
\alias{create_survey_circumcision_dhs}
\title{Create male circumcision outcomes dataset from DHS}
\usage{
create_survey_circumcision_dhs(surveys)
}
\arguments{
\item{surveys}{data.frame of surveys, returned by \code{create_surveys_dhs()}.}
}
\value{
data.frame consisting of survey ID, individual ID and male circumcision
outcomes. See details.
}
\description{
Create male circumcision outcomes dataset from DHS
}
\details{
The following fields are extracted:
\itemize{
\item survey_id
\item individual_id
\item circumcised
\item circ_age
\item circ_where
\item circ_who
}
}
\examples{

surveys <- create_surveys_dhs("MWI")
circ <- create_circumcision_dhs(surveys)

}
