% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{create_surveys_dhs}
\alias{create_surveys_dhs}
\title{Create surveys dataset from DHS API}
\usage{
create_surveys_dhs(
  iso3,
  survey_type = c("DHS", "AIS", "MIS"),
  survey_characteristics = 23
)
}
\arguments{
\item{iso3}{Three letter ISO3 country code.}

\item{survey_characteristics}{DHS survey characteristic IDs to filter on See \code{?rdhs::dhs_survey_characteristics}.}

\item{survey_types}{DHS survey types to access. See \code{?rdhs::dhs_surveys}.}
}
\value{
A data frame containing the response from the \emph{dhs_surveys} API endpoint
and the \code{survey_id} and \code{survey_mid_calendar_quarter}.
}
\description{
Construct a surveys dataset from DHS API. Usess \code{rdhs} to identify the DHS
country code from the ISO3, selects relevant surveys, then constructs the
\code{survey_id} and \code{survey_mid_calendar_quarter}.
}
\examples{
create_surveys_dhs("MWI")

}
