% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{generate_shipp_tool}
\alias{generate_shipp_tool}
\title{Generate SHIPP tool}
\usage{
generate_shipp_tool(output, pjnz, path = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{output}{Path to naomi outputs}

\item{pjnz}{Path to input PJNZ file}

\item{path}{Path to save output file}
}
\value{
Path to output file and metadata for file
}
\description{
Generate SHIPP tool
}
