% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{hintr_inputs_ready}
\alias{hintr_inputs_ready}
\title{Prepare output from hintr debug rds for debugging}
\usage{
hintr_inputs_ready(jobid, root = ".")
}
\arguments{
\item{jobid}{The issue ID, the name of the folder in sharepoint}

\item{root}{The debug root dir}
}
\value{
Path to local debug
}
\description{
Prepare output from hintr debug rds for debugging
}
