% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{naomi_debug}
\alias{naomi_debug}
\title{Download debug from server and upload into sharepoint}
\usage{
naomi_debug(
  id,
  issue_id,
  dest_folder = "Shared Documents/2023_debug",
  server = NULL
)
}
\arguments{
\item{id}{The model fit or calibrate ID to download debug for}

\item{issue_id}{The issue ID, the name of the folder to create in sharepoint}

\item{dest_folder}{The root destination folder in sharepoint}

\item{server}{The folder to download debug from, defaults to production
server}
}
\value{
Path to local debug
}
\description{
Download debug from server and upload into sharepoint
}
