% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survey_dhs.R
\name{plot_survey_coordinate_check}
\alias{plot_survey_coordinate_check}
\title{Summary plot of survey cluster coordinates outside boundaries}
\usage{
plot_survey_coordinate_check(
  survey_clusters,
  survey_region_boundaries,
  survey_region_areas
)
}
\arguments{
\item{survey_clusters}{Survey clusteres dataset}
}
\value{
A list of grobs, one for each survey.
}
\description{
Summary plot of survey cluster coordinates outside boundaries
}
