% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pjnz.R
\name{read_shiny90_country}
\alias{read_shiny90_country}
\title{Read country from .zip.shiny90 file}
\usage{
read_shiny90_country(shiny90_zip)
}
\arguments{
\item{shiny90_zip}{path to .shiny90 export file}
}
\value{
Shiny90 country / region name.
}
\description{
Read country from .zip.shiny90 file
}
