% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_calculate_incidence_male}
\alias{shipp_calculate_incidence_male}
\title{Calculate incidence in all behavioural groups}
\usage{
shipp_calculate_incidence_male(
  naomi_output,
  options,
  iso,
  male_srb,
  male_logit_prevalence,
  survey_year = 2018,
  consensus_est,
  goals,
  kp_wb
)
}
\arguments{
\item{naomi_output}{Naomi output.}

\item{options}{Naomi model options.}

\item{iso}{ISO3 country code.}

\item{male_srb}{MSM and PWID adjusted estimated of male SRB groups generated by \code{shipp_adjust_sexbehav_msm_pwid()}.}

\item{male_logit_prevalence}{Risk adjusted estimates of male HIV prevalence in sexual risk groups generated by \code{shipp_calculate_prevalence_male()}.}

\item{survey_year}{Survey year to sample from the SAE model. Default is 2018. Survey year should be updated to most current household survey in the country - for countries without recent household surveys, leave at 2018 - the spatiotemporal
model of sexual behaviour fitted to all countries has the most data for in roughly 2018.}

\item{consensus_est}{Source to scale PSE + new infections.}

\item{goals}{National PSE and new infection estimates from Goals.}

\item{kp_wb}{National PSE and new infection estimates from KP Workbook read from pjnz.}
}
\value{
Wide format output required for the SHIPP workbook
}
\description{
Calculate incidence in all behavioural groups
}
