% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipp.R
\name{shipp_disaggregate_fsw}
\alias{shipp_disaggregate_fsw}
\title{Dissagreggate admin1 FSW proportions from Oli's KP model to 5-age groups}
\usage{
shipp_disaggregate_fsw(outputs, iso, naomi_pop, consensus_est, goals, kp_wb)
}
\arguments{
\item{outputs}{Naomi output.}

\item{iso}{Country iso3.}

\item{naomi_pop}{Naomi population estimates for current time point.}

\item{consensus_est}{Source to scale PSE + new infections.}

\item{goals}{National PSE and new infection estimates from Goals.}

\item{kp_wb}{National PSE and new infection estimates from KP Workbook read from pjnz.}
}
\value{
District level FSW estimates by 5-year age bands for ages 15-49.
}
\description{
Dissagreggate admin1 FSW proportions from Oli's KP model to 5-age groups
}
\keyword{internal}
